test(toolloop): split default tool loop system tests
